refactor: reorganize Tooltip component structure and clean up themeConfig
- Split the SplitTag component into its own file for better modularity.
- Updated the Tooltip component to streamline imports and remove unnecessary props.
- Cleaned up the themeConfig by removing unused tooltip styles, enhancing maintainability.
- Adjusted utility functions for Tooltip to ensure consistent class name generation and improved readability.
