+++ lib/main.ts
@@ -3,6 +3,7 @@ import '../src/index.css';
 export { default as Button } from './components/Button/Button';
 export { default as Menu } from './components/Menu/Menu';
 export { default as Checkbox } from './components/common/Checkbox';
-export { default as Tag, SplitTag } from './components/Tag/Tag';
+export { default as Tag } from './components/Tag/Tag';
+export { default as SplitTag } from './components/Tag/SplitTag';
 export { default as Tooltip } from './components/Tooltip/Tooltip';
 export { Tabs, TabsList, TabsTrigger, TabsContent } from './components/Tabs';
